requirements.txt
python-telegram-bot>=20.7
opencv-python-headless>=4.8.0
pytesseract>=0.3.10
Pillow>=10.2.0
schedule>=1.2.0
python-dotenv>=1.0.0
aiofiles>=23.2.0
numpy>=1.24.0
regex>=2023.10.0
gunicorn>=21.2.0
psycopg2-binary>=2.9.0
flask>=3.0.0 